requirements.txt
# 核心依赖
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.10,<0.1.0
langchain-text-splitters>=0.0.1,<0.1.0
langchain-chroma>=0.1.0,<0.2.0
langchain-openai>=0.1.0,<0.2.0

# OpenAI相关 - 使用兼容版本
openai>=1.10.0,<2.0.0  # 这是langchain-openai要求的版本

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
mysql-connector-python==8.2.0
redis==5.0.1
sqlalchemy==2.0.23

# 数据处理
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0

# 文本处理
pydantic==2.5.0
pydantic-settings==2.1.0

# 音频处理
speechrecognition==3.10.0
pydub==0.25.1

# 图像处理
Pillow==10.1.0
pytesseract==0.3.10

# 其他工具
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2